// src/types/modules.ts
export type CovidZoneId = "A" | "B" | "C" | "D" | "E" | "Unknown";
export type CovidZones = Record<Exclude<CovidZoneId, "Unknown">, CovidZone>;

export type AcadYear = string; // E.g. "2016/2017"
export type ClassNo = string; // E.g. "1", "A"
export type DayText = string; // E.g. "Monday", "Tuesday"
export type Department = string;
export type StartTime = string; // E.g. "1400"
export type EndTime = string; // E.g. "1500"
export type Faculty = string;
export type LessonType = string; // E.g. "Lecture", "Tutorial"
export type LessonTime = StartTime | EndTime;
export type ModuleCode = string; // E.g. "CS3216"
export type ModuleTitle = string;
export type Semester = number; // E.g. 1/2/3/4. 3 and 4 means special sem i and ii.
export type Workload = string | number[];
// ...
export type PrereqTree = string | { and?: PrereqTree[]; or?: PrereqTree[] };
// ...
export type RawLesson = Readonly<{
    classNo: ClassNo;
    day: DayText;
    endTime: EndTime;
    lessonType: LessonType;
    startTime: StartTime;
    venue: Venue;
    weeks: Weeks;
    size: number;
    covidZone: CovidZoneId;
}>;

// Semester-specific information of a module.
export type SemesterData = {
    semester: Semester;
    timetable: RawLesson[];

    // Aggregated from timetable
    covidZones: CovidZoneId[];

    // Exam
    examDate?: string;
    examDuration?: number;
};

export type NUSModuleAttributes = Partial<{
    year: boolean; // Year long
    su: boolean; // Can S/U (undergraduate)
    grsu: boolean; // Can S/U (graduate)
    ssgf: boolean; // SkillsFuture Funded
    sfs: boolean; // SkillsFuture series
    lab: boolean; // Lab based
    ism: boolean; // Independent study
    urop: boolean; // Undergraduate Research Opportunities Program
    fyp: boolean; // Honours / Final Year Project
    mpes1: boolean; // Included in Semester 1's Module Planning Exercise
    mpes2: boolean; // Included in Semester 2's Module Planning Exercise
}>;

// Information for a module for a particular academic year.
export type Module = {
    acadYear: AcadYear;

    // Basic info
    moduleCode: ModuleCode;
    title: ModuleTitle;

    // Additional info
    description?: string;
    moduleCredit: string;
    department: Department;
    faculty: Faculty;
    workload?: Workload;
    aliases?: ModuleCode[];
    attributes?: NUSModuleAttributes;

    // Requsites
    prerequisite?: string;
    corequisite?: string;
    preclusion?: string;

    // Semester data
    semesterData: SemesterData[];

    // Requisite tree
    prereqTree?: PrereqTree;
    fulfillRequirements?: ModuleCode[];
};